backend/requirements.txt
aiogram>=3.3,<4
fastapi>=0.109
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
pydantic-settings>=2.0
apscheduler>=3.10,<4
python-dateutil>=2.8
